+++ Sources/ToastUI/DefaultToastViewStyle.swift
@@ -12,8 +12,8 @@ struct AnyToastViewStyle: ToastViewStyle {
   private let _makeBody: (Configuration) -> AnyView
 
   /// Creates a ``AnyToastViewStyle``.
-  init<Style>(_ style: Style) where Style: ToastViewStyle {
-    _makeBody = { configuration in
+  init(_ style: some ToastViewStyle) {
+    self._makeBody = { configuration in
       AnyView(style.makeBody(configuration: configuration))
     }
   }
@@ -29,7 +29,7 @@ struct AnyToastViewStyle: ToastViewStyle {
 /// The default style of ``ToastView``.
 ///
 /// Use this style to customize ``ToastView``.
-@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, *)
+@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, visionOS 1.0, *)
 public struct DefaultToastViewStyle: ToastViewStyle {
   /// Creates a ``DefaultToastViewStyle``.
   public init() {}
@@ -48,37 +48,27 @@ public struct DefaultToastViewStyle: ToastViewStyle {
   }
 }
 
-@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, *)
-public extension ToastViewStyle where Self == DefaultToastViewStyle {
+@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, visionOS 1.0, *)
+extension ToastViewStyle where Self == DefaultToastViewStyle {
   /// The default style of ``ToastView``.
   ///
   /// Use this style to customize ``ToastView``.
-  static var `default`: Self {
+  public static var `default`: Self {
     .init()
   }
 }
 
 /// A ``ToastView`` represents as an indeterminate circular progress indicator,
 /// also informally known as a spinner.
-@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, *)
+@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, visionOS 1.0, *)
 public struct IndeterminateProgressToastViewStyle: ToastViewStyle {
-  /// Creates an ``IndeterminateProgressToastViewStyle``.
-  public init() {}
-
-  /// Creates a view representing the body of a ``ToastView``.
-  ///
-  /// - Parameter configuration: The properties of the ``ToastView`` being created.
-  public func makeBody(configuration: Configuration) -> some View {
-    IndeterminateProgressToastView(configuration: configuration)
-  }
-
   struct IndeterminateProgressToastView: View {
     let configuration: Configuration
 
-    @ScaledMetric private var iconSize: Double = 36
-    @ScaledMetric private var strokeSize: Double = 3
+    @ScaledMetric private var iconSize = 36.0
+    @ScaledMetric private var strokeSize = 3.0
 
-    @State private var animating: Bool = false
+    @State private var animating = false
 
     var body: some View {
       DefaultToastView {
@@ -96,7 +86,10 @@ public struct IndeterminateProgressToastViewStyle: ToastViewStyle {
           .frame(width: iconSize, height: iconSize)
           .rotationEffect(.degrees(-90))
           .rotationEffect(.degrees(animating ? 360 : 0))
-          .animation(.linear(duration: 1.0).repeatForever(autoreverses: false), value: animating)
+          .animation(
+            .linear(duration: 1.0).repeatForever(autoreverses: false),
+            value: animating
+          )
           .onAppear {
             // fix #2
             DispatchQueue.main.async {
@@ -112,13 +105,23 @@ public struct IndeterminateProgressToastViewStyle: ToastViewStyle {
       }
     }
   }
+
+  /// Creates an ``IndeterminateProgressToastViewStyle``.
+  public init() {}
+
+  /// Creates a view representing the body of a ``ToastView``.
+  ///
+  /// - Parameter configuration: The properties of the ``ToastView`` being created.
+  public func makeBody(configuration: Configuration) -> some View {
+    IndeterminateProgressToastView(configuration: configuration)
+  }
 }
 
-@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, *)
-public extension ToastViewStyle where Self == IndeterminateProgressToastViewStyle {
+@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, visionOS 1.0, *)
+extension ToastViewStyle where Self == IndeterminateProgressToastViewStyle {
   /// A ``ToastView`` represents as an indeterminate circular progress indicator,
   /// also informally known as a spinner.
-  static var indeterminate: Self {
+  public static var indeterminate: Self {
     .init()
   }
 }
@@ -126,39 +129,18 @@ public extension ToastViewStyle where Self == IndeterminateProgressToastViewStyl
 /// A ``ToastView`` represents as a determinate circular progress indicator.
 /// This style is visually similar to the ``IndeterminateProgressToastViewStyle``,
 /// but shows determinate progress instead.
-@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, *)
+@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, visionOS 1.0, *)
 public struct DeterminateProgressToastViewStyle<Value>: ToastViewStyle
-where Value: BinaryFloatingPoint {
-  @Binding private var value: Value
-  @Binding private var total: Value
-
-  /// Creates a ``DeterminateProgressToastViewStyle``.
-  ///
-  /// - Parameters:
-  ///   - value: The completed amount of the task, ranging from `0.0` to `total`.
-  ///   - total: The fully completed amount of the task, meaning the task is completed if
-  ///            `value` equals `total`. Default value is `1.0`.
-  public init(value: Binding<Value>, total: Binding<Value> = .constant(1.0)) {
-    _value = value
-    _total = total
-  }
-
-  /// Creates a view representing the body of a ``ToastView``.
-  ///
-  /// - Parameter configuration: The properties of the ``ToastView`` being created.
-  public func makeBody(configuration: Configuration) -> some View {
-    DeterminateProgressToastView(configuration: configuration, value: $value, total: $total)
-  }
-
+  where Value: BinaryFloatingPoint {
   struct DeterminateProgressToastView: View {
     let configuration: Configuration
     @Binding var value: Value
     @Binding var total: Value
 
-    @ScaledMetric private var iconSize: Double = 36
-    @ScaledMetric private var strokeSize: Double = 3
+    @ScaledMetric private var iconSize = 36.0
+    @ScaledMetric private var strokeSize = 3.0
 
-    @State private var isAnimating: Bool = false
+    @State private var isAnimating = false
 
     var body: some View {
       DefaultToastView {
@@ -185,10 +167,31 @@ where Value: BinaryFloatingPoint {
       }
     }
   }
+
+  @Binding private var value: Value
+  @Binding private var total: Value
+
+  /// Creates a ``DeterminateProgressToastViewStyle``.
+  ///
+  /// - Parameters:
+  ///   - value: The completed amount of the task, ranging from `0.0` to `total`.
+  ///   - total: The fully completed amount of the task, meaning the task is completed if
+  ///            `value` equals `total`. Default value is `1.0`.
+  public init(value: Binding<Value>, total: Binding<Value> = .constant(1.0)) {
+    _value = value
+    _total = total
+  }
+
+  /// Creates a view representing the body of a ``ToastView``.
+  ///
+  /// - Parameter configuration: The properties of the ``ToastView`` being created.
+  public func makeBody(configuration: Configuration) -> some View {
+    DeterminateProgressToastView(configuration: configuration, value: $value, total: $total)
+  }
 }
 
-@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, *)
-public extension ToastViewStyle {
+@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, visionOS 1.0, *)
+extension ToastViewStyle {
   /// A ``ToastView`` represents as a determinate circular progress indicator.
   /// This style is similar to the ``IndeterminateProgressToastViewStyle``,
   /// but shows determinate progress instead.
@@ -197,34 +200,18 @@ public extension ToastViewStyle {
   ///   - value: The completed amount of the task, ranging from `0.0` to `total`.
   ///   - total: The fully completed amount of the task, meaning the task is completed if
   ///            `value` equals `total`. Default value is `1.0`.
-  static func determinate<Value>(
+  public static func determinate<Value>(
     value: Binding<Value>,
     total: Binding<Value> = .constant(1.0)
-  ) -> Self where Value: BinaryFloatingPoint, Self == DeterminateProgressToastViewStyle<Value> {
+  )
+    -> Self where Value: BinaryFloatingPoint, Self == DeterminateProgressToastViewStyle<Value> {
     .init(value: value, total: total)
   }
 }
 
 /// A ``ToastView`` represents as a toast consists of an icon and a headline text label.
-@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, *)
+@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, visionOS 1.0, *)
 public struct IconToastViewStyle<Content>: ToastViewStyle where Content: View {
-  private var content: Content
-
-  /// Creates an ``IconToastViewStyle``.
-  ///
-  /// - Parameters:
-  ///   - content: The content of the icon.
-  public init(content: () -> Content) {
-    self.content = content()
-  }
-
-  /// Creates a view representing the body of a ``ToastView``.
-  ///
-  /// - Parameter configuration: The properties of the ``ToastView`` being created.
-  public func makeBody(configuration: Configuration) -> some View {
-    IconToastView(configuration: configuration, content: content)
-  }
-
   struct IconToastView: View {
     let configuration: Configuration
     let content: Content
@@ -244,22 +231,39 @@ public struct IconToastViewStyle<Content>: ToastViewStyle where Content: View {
       }
     }
   }
+
+  private var content: Content
+
+  /// Creates an ``IconToastViewStyle``.
+  ///
+  /// - Parameters:
+  ///   - content: The content of the icon.
+  public init(content: () -> Content) {
+    self.content = content()
+  }
+
+  /// Creates a view representing the body of a ``ToastView``.
+  ///
+  /// - Parameter configuration: The properties of the ``ToastView`` being created.
+  public func makeBody(configuration: Configuration) -> some View {
+    IconToastView(configuration: configuration, content: content)
+  }
 }
 
-@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, *)
-public extension ToastViewStyle {
+@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, visionOS 1.0, *)
+extension ToastViewStyle {
   /// A ``ToastView`` represents as a toast consists of an icon and a headline text label.
   ///
   /// - Parameters:
   ///   - content: The content of the icon.
-  static func icon<Content>(content: () -> Content) -> Self
-  where Self == IconToastViewStyle<Content> {
+  public static func icon<Content>(content: () -> Content) -> Self
+    where Self == IconToastViewStyle<Content> {
     .init(content: content)
   }
 }
 
 /// A ``ToastView`` represents as a success toast.
-@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, *)
+@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, visionOS 1.0, *)
 public struct SuccessToastViewStyle: ToastViewStyle {
   /// Creates a ``SuccessToastViewStyle``.
   public init() {}
@@ -276,16 +280,16 @@ public struct SuccessToastViewStyle: ToastViewStyle {
   }
 }
 
-@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, *)
-public extension ToastViewStyle where Self == SuccessToastViewStyle {
+@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, visionOS 1.0, *)
+extension ToastViewStyle where Self == SuccessToastViewStyle {
   /// A ``ToastView`` represents as a success toast.
-  static var success: Self {
+  public static var success: Self {
     .init()
   }
 }
 
 /// A ``ToastView`` represents as a failure toast.
-@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, *)
+@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, visionOS 1.0, *)
 public struct FailureToastViewStyle: ToastViewStyle {
   /// Creates a ``FailureToastViewStyle``.
   public init() {}
@@ -302,16 +306,16 @@ public struct FailureToastViewStyle: ToastViewStyle {
   }
 }
 
-@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, *)
-public extension ToastViewStyle where Self == FailureToastViewStyle {
+@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, visionOS 1.0, *)
+extension ToastViewStyle where Self == FailureToastViewStyle {
   /// A ``ToastView`` represents as a failure toast.
-  static var failure: Self {
+  public static var failure: Self {
     .init()
   }
 }
 
 /// A ``ToastView`` represents as a warning toast.
-@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, *)
+@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, visionOS 1.0, *)
 public struct WarningToastViewStyle: ToastViewStyle {
   /// Creates a ``WarningToastViewStyle``.
   public init() {}
@@ -328,16 +332,16 @@ public struct WarningToastViewStyle: ToastViewStyle {
   }
 }
 
-@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, *)
-public extension ToastViewStyle where Self == WarningToastViewStyle {
+@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, visionOS 1.0, *)
+extension ToastViewStyle where Self == WarningToastViewStyle {
   /// A ``ToastView`` represents as a warning toast.
-  static var warning: Self {
+  public static var warning: Self {
     .init()
   }
 }
 
 /// A ``ToastView`` represents as an information toast.
-@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, *)
+@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, visionOS 1.0, *)
 public struct InformationToastViewStyle: ToastViewStyle {
   /// Creates an ``InformationToastViewStyle``.
   public init() {}
@@ -354,10 +358,10 @@ public struct InformationToastViewStyle: ToastViewStyle {
   }
 }
 
-@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, *)
-public extension ToastViewStyle where Self == InformationToastViewStyle {
+@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, visionOS 1.0, *)
+extension ToastViewStyle where Self == InformationToastViewStyle {
   /// A ``ToastView`` represents as an information toast.
-  static var information: Self {
+  public static var information: Self {
     .init()
   }
 }