+++ Sources/ToastUI/ToastViewStyle.swift
@@ -12,7 +12,7 @@ import SwiftUI
 ///
 /// To configure the current ``ToastViewStyle`` for a view hiearchy, use the
 /// ``ToastView/toastViewStyle(_:)`` modifier.
-@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, *)
+@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, visionOS 1.0, *)
 public protocol ToastViewStyle {
   /// A view that represents the body of a ``ToastView``.
   associatedtype Body: View
@@ -27,36 +27,36 @@ public protocol ToastViewStyle {
 }
 
 /// The properties of a ``ToastView``.
-@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, *)
+@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, visionOS 1.0, *)
 public struct ToastViewStyleConfiguration {
   /// A type-erased background of a ``ToastView``.
   public struct Background: View {
-    init<Content>(content: Content) where Content: View {
-      body = AnyView(content)
-    }
-
     /// The body of this view.
     public var body: AnyView
+
+    init(content: some View) {
+      self.body = AnyView(content)
+    }
   }
 
   /// A type-erased label of a ``ToastView``.
   public struct Label: View {
-    init<Content>(content: Content) where Content: View {
-      body = AnyView(content)
-    }
-
     /// The body of this view.
     public var body: AnyView
+
+    init(content: some View) {
+      self.body = AnyView(content)
+    }
   }
 
   /// A type-erased content of a ``ToastView``.
   public struct Content: View {
-    init<Content>(content: Content) where Content: View {
-      body = AnyView(content)
-    }
-
     /// The body of this view.
     public var body: AnyView
+
+    init(content: some View) {
+      self.body = AnyView(content)
+    }
   }
 
   /// A view describes the background of a ``ToastView``.